- Rewrote the console to reflect the following user journey:
searchByForm():
User decides to search by form. A menu of forms is displayed.
Next, the user chooses a form. A list of titles pertaining to that form
is displayed. This is done via accessing the formMap at key = chosen
form.
Next, the user picks a title pressing a number (i.e., index in
ArrayList). The poem is displayed.
Next, the console asks: "Do you want to add this poem to your anthology?
Yes/No.
If yes, we add the poem to a private instance variable named
userAnthology (ArrayLists<Poem>).
Either way (yes or no), console asks: "Are you done with picking?"
If yes, write out .txt file with all poems in userAnthology. If No, go
back to initial menu.
Use index from map to choose which poems to send to write method

This applies to form, theme, and author

- As a result, made changes in DataBaseManageClass and Poem Class. Added
further comments to lines of code + JavaDoc comments.